chore: move helpers to test
