+++ color.go
@@ -1,7 +1,5 @@
 package lipgloss
 
-import "image/color"
-
 // TerminalColor is a color intended to be rendered in the terminal.
 type TerminalColor interface {
 	colorType()
@@ -34,10 +32,9 @@ func (Color) colorType() {}
 //
 // Example usage:
 //
-//    // These two statements are equivalent.
-//    colorA := lipgloss.ANSIColor(21)
-//    colorB := lipgloss.Color("21")
-//
+//	// These two statements are equivalent.
+//	colorA := lipgloss.ANSIColor(21)
+//	colorB := lipgloss.Color("21")
 type ANSIColor uint
 
 func (ANSIColor) colorType() {}
@@ -75,48 +72,3 @@ type CompleteAdaptiveColor struct {
 }
 
 func (CompleteAdaptiveColor) colorType() {}
-
-// hexToColor translates a hex color string (#RRGGBB or #RGB) into a color.RGB,
-// which satisfies the color.Color interface. If an invalid string is passed
-// black with 100% opacity will be returned: or, in hex format, 0x000000FF.
-func hexToColor(hex string) (c color.RGBA) {
-	c.A = 0xFF
-
-	if hex == "" || hex[0] != '#' {
-		return c
-	}
-
-	const (
-		fullFormat  = 7 // #RRGGBB
-		shortFormat = 4 // #RGB
-	)
-
-	switch len(hex) {
-	case fullFormat:
-		const offset = 4
-		c.R = hexToByte(hex[1])<<offset + hexToByte(hex[2])
-		c.G = hexToByte(hex[3])<<offset + hexToByte(hex[4])
-		c.B = hexToByte(hex[5])<<offset + hexToByte(hex[6])
-	case shortFormat:
-		const offset = 0x11
-		c.R = hexToByte(hex[1]) * offset
-		c.G = hexToByte(hex[2]) * offset
-		c.B = hexToByte(hex[3]) * offset
-	}
-
-	return c
-}
-
-func hexToByte(b byte) byte {
-	const offset = 10
-	switch {
-	case b >= '0' && b <= '9':
-		return b - '0'
-	case b >= 'a' && b <= 'f':
-		return b - 'a' + offset
-	case b >= 'A' && b <= 'F':
-		return b - 'A' + offset
-	}
-	// Invalid, but just return 0.
-	return 0
-}
